% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupInlabruInputs.R
\name{create_input_list_temporal_noCatalogue}
\alias{create_input_list_temporal_noCatalogue}
\title{Function to create a default input list for the ETAS Hawkes temporal model where no catalogue is specified in the input file}
\usage{
create_input_list_temporal_noCatalogue(input_path)
}
\arguments{
\item{input_path}{Input file and path as a string}
}
\value{
The formatted input.list with the elements required for the temporal Hawkes model
}
\description{
Function to create a default input list for the ETAS Hawkes temporal model where no catalogue is specified in the input file
}
\examples{
create_input_list_temporal_noCatalogue(
  system.file("extdata", "user_input_synthetic_noCatalogue.txt", package = "ETAS.inlabru"))
}
