% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupInlabruInputs.R
\name{create_input_list_temporal_withCatalogue}
\alias{create_input_list_temporal_withCatalogue}
\title{Function to create a default input file for the ETAS Hawkes temporal model where a catalogue is specified in the input file.}
\usage{
create_input_list_temporal_withCatalogue(input_path)
}
\arguments{
\item{input_path}{path of the \code{txt} file containing experiment's information}
}
\value{
The formatted input.list with the elements required for the temporal Hawkes model
}
\description{
Function to create a default input file for the ETAS Hawkes temporal model where a catalogue is specified in the input file.
}
