% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaTransformations.R
\name{exp.t}
\alias{exp.t}
\title{Copula transformation from a standard Normal distribution to an Exponential distribution}
\usage{
exp.t(x, r)
}
\arguments{
\item{x}{values from a standard Normal distribution, \code{vector}.}

\item{r}{rate of the exponential distribution, \code{scalar}.}
}
\value{
values from an Exponential distribution with rate \code{r}, \code{vector} same length as \code{x}.
}
\description{
Copula transformation from a standard Normal distribution to an Exponential distribution
}
