% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaTransformations.R
\name{exp_t}
\alias{exp_t}
\title{Copula transformation from a standard Normal distribution to an Exponential distribution}
\usage{
exp_t(x, r)
}
\arguments{
\item{x}{values from a standard Normal distribution, \code{vector}.}

\item{r}{rate of the exponential distribution, \code{scalar}.}
}
\value{
values from an Exponential distribution with rate \code{r}, \code{vector} same length as \code{x}.
}
\description{
Copula transformation from a standard Normal distribution to an Exponential distribution
}
\seealso{
Other copula-transformations: 
\code{\link{gamma_t}()},
\code{\link{inv_exp_t}()},
\code{\link{inv_gamma_t}()},
\code{\link{inv_loggaus_t}()},
\code{\link{inv_unif_t}()},
\code{\link{loggaus_t}()},
\code{\link{unif_t}()}
}
\concept{copula-transformations}
