% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticCatalogues.R
\name{generate.temporal.ETAS.synthetic}
\alias{generate.temporal.ETAS.synthetic}
\title{Generates a sythetic catalogue using the ETAS model}
\usage{
\method{generate}{temporal.ETAS.synthetic}(theta, beta.p, M0, T1, T2, Ht = NULL, ncore = 1)
}
\arguments{
\item{theta}{ETAS parameters \code{list(mu=mu, K=K, alpha=alpha, c=c, p=p)}.}

\item{beta.p}{Slope of GR relation: beta = b ln(10).}

\item{M0}{The minimum magnitude in the synthetic catalogue.}

\item{T1}{The start time for the synthetic catalogue \link{days}.}

\item{T2}{The end time for the synthetic catalogue \link{days}.}

\item{Ht}{A catalogue history to impose on the synthetic sequence.}

\item{ncore}{Integer number of compute cores to use.}
}
\value{
A data.frame of the temporal catalogue with columns \verb{[t_i, M_i, gen_i]}
where, \code{t_i} are the times,  \code{M_i} the magnitudes, \code{gen_i} includes information about the generation number
}
\description{
Generates a sythetic catalogue using the ETAS model
}
\examples{
## EXAMPLE 1: Generate a 1000 day synthetic ETAS catalogue

generate.temporal.ETAS.synthetic( theta= list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08), beta.p=log(10), M0=2.5, T1=0, T2=1000 )


## EXAMPLE 2: To generate a 1000 day catalogue including a M6.7 event on day 500

Ht <- data.frame(ts=c(500), magnitudes=c(6.7))
generate.temporal.ETAS.synthetic( theta= list(mu=0.1, K=0.089, alpha=2.29, c=0.11, p=1.08), beta.p=log(10), M0=2.5, T1=0, T2=1000, Ht=Ht )
}
