% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticCatalogues.R
\name{generate_temporal_ETAS_synthetic}
\alias{generate_temporal_ETAS_synthetic}
\title{Generates a synthetic catalogue using the ETAS model}
\usage{
generate_temporal_ETAS_synthetic(
  theta,
  beta.p,
  M0,
  T1,
  T2,
  Ht = NULL,
  ncore = 1,
  format = "list"
)
}
\arguments{
\item{theta}{ETAS parameters \code{list(mu=mu, K=K, alpha=alpha, c=c, p=p)}.}

\item{beta.p}{Slope of GR relation: beta = b ln(10).}

\item{M0}{The minimum magnitude in the synthetic catalogue.}

\item{T1}{The start time for the synthetic catalogue \verb{[days]}.}

\item{T2}{The end time for the synthetic catalogue \verb{[days]}.}

\item{Ht}{A catalogue history to impose on the synthetic sequence.}

\item{ncore}{Integer number of compute cores to use.}

\item{format}{If "list", return a list of data.frame objects, one for each generation.
If "df", return a data.frame where the list element have been joined into a
single data.frame, and ordered by \code{ts}}
}
\value{
A list of data.frame objects of the temporal catalogue with
columns \verb{[ts, magnitudes, gen]} where, \code{ts} are the times \code{t_i},
\code{magnitudes} the magnitudes \code{M_i}, and \code{gen} includes information about the
generation number, \code{gen_i}.

To join into an time-ordered data.frame,

\if{html}{\out{<div class="sourceCode">}}\preformatted{cata <- generate_temporal_ETAS_synthetic(...)
cata <- dplyr::bind_rows(cata)
cata <- dplyr::arrange(cata, ts)
}\if{html}{\out{</div>}}

or

\if{html}{\out{<div class="sourceCode">}}\preformatted{cata <- generate_temporal_ETAS_synthetic(..., format = "df")
}\if{html}{\out{</div>}}
}
\description{
Generates a synthetic catalogue using the ETAS model
}
\examples{
## EXAMPLE 1: Generate a 1000 day synthetic ETAS catalogue

generate_temporal_ETAS_synthetic(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000
)


## EXAMPLE 2: To generate a 1000 day catalogue including a M6.7 event on day 500

Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
generate_temporal_ETAS_synthetic(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)

}
