% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysingPosteriors.R
\name{get_posterior_N}
\alias{get_posterior_N}
\title{Title}
\usage{
get_posterior_N(input.list)
}
\arguments{
\item{input.list}{Which has combined the input file (for link functions) and bru output (for marginals)}
}
\description{
Title
}
