% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysingPosteriors.R
\name{get_posterior_param}
\alias{get_posterior_param}
\title{Generate summary information on the fitted ETAS model}
\usage{
get_posterior_param(input.list)
}
\arguments{
\item{input.list}{Which has combined the input file (for link functions) and bru output (for marginals)}
}
\value{
Data frame summary and summary plot
}
\description{
Generate summary information on the fitted ETAS model
}
