% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETAS.triggering.functions.R
\name{gt.2}
\alias{gt.2}
\title{Time triggering function - used by Inlabru
MN: TODO Cross-reference the paper}
\usage{
gt.2(th, t, ti, mi, M0)
}
\arguments{
\item{M0}{Minimum magnitude threshold}
}
\description{
Time triggering function - used by Inlabru
MN: TODO Cross-reference the paper
}
