% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HawkesProcess.R
\name{Hawkes.bru}
\alias{Hawkes.bru}
\title{Function to fit Hawkes process model}
\usage{
Hawkes.bru(
  sample.s,
  M0,
  T1,
  T2,
  link.functions = NULL,
  coef.t.,
  delta.t.,
  N.max.,
  bru.opt
)
}
\arguments{
\item{M0}{Minimum magnitude threshold}

\item{T1}{Start of temporal model domain.}

\item{T2}{End of temporal model domain.}

\item{coef.t.}{TimeBinning parameter:}

\item{delta.t.}{TimeBinning parameter:}

\item{N.max.}{TimeBinning parameter: Number of bins}

\item{bru.opt}{}
}
\description{
Function to fit Hawkes process model
}
