% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HawkesProcess.R
\name{Hawkes.bru2}
\alias{Hawkes.bru2}
\title{Function to fit Hawkes process model}
\usage{
Hawkes.bru2(
  sample.s,
  M0,
  T1,
  T2,
  link.functions = NULL,
  coef.t.,
  delta.t.,
  N.max.,
  bru.opt
)
}
\arguments{
\item{M0}{Minimum magnitude threshold}

\item{T1}{Start of temporal model domain \link{days}.}

\item{T2}{End of temporal model domain \link{days}.}

\item{coef.t.}{TimeBinning parameter:}

\item{delta.t.}{TimeBinning parameter:}

\item{N.max.}{TimeBinning parameter: Number of bins}

\item{bru.opt}{Runtime options for inlabru: See https://inlabru-org.github.io/inlabru/reference/bru_call_options.html}
}
\value{
The inlabru Hawkes process results data.frame
}
\description{
Function to fit Hawkes process model
}
