% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.horus.R
\docType{data}
\name{horus}
\alias{horus}
\title{HORUS Ita Catalogue}
\format{
Original file format:
\describe{
\item{\code{Year}:}{Origin Time (OT) year}
\item{\code{Mo}:}{OT month}
\item{\code{Da}:}{OT day}
\item{\code{Ho}:}{OT hour}
\item{\code{Mi}}{OT minute}
\item{\code{Se}}{OT seconds and fractions}
\item{\code{Lat}}{epicenter N latitude (in degrees)}
\item{\code{Lon}}{epicenter E longitude (in degrees)}
\item{\code{Depth}}{hypocenter depth (in km)}
\item{\code{Mw}}{true or proxy moment magnitude}
\item{\code{sigMw}}{moment magnitude uncertainty}
\item{\code{Geo.Ita}}{\code{"*"} indicates that the epicenter is within the Italian
mainland territory, otherwise \code{" "}}
\item{\code{Geo.CPTI15}}{\code{"*"} indicates that the epicenter is within the spatial
window of the CPTI15 catalog (Rovida et al., 2020, Bull Earth Eng,
doi: 10.1007/s10518-020-00818-y)}
\item{\code{Ev..type}}{\code{"x"} indicates that the event is not an earthquake
(e.g. explosion, eruption, landslide, ...) (only since May 1st 2012)}
\item{\code{Iside.n.}}{ISIDe id number (only since April 16th 2005)}
}
ETAS.inlabru format:
\describe{
\item{\code{time_String}:}{Combination of original \code{Year}, \code{Mo}, \code{Da}, \code{Mi}, and \code{Se}}
\item{\code{lon}:}{Original \code{Lon}}
\item{\code{lat}:}{Original \code{Lat}}
\item{\code{depth}}{Original \code{Depth}}
\item{\code{M}:}{Original \code{Mw}}
}
}
\source{
\url{http://horus.bo.ingv.it/}
}
\usage{
data(horus, package = "ETAS.inlabru")
}
\description{
The HOmogenized instRUmental Seismic catalog (HORUS) of Italy from 1960 to present
}
\details{
The original entire HORUS catalog is provided as a single tab separated ascii file:
in \code{HORUS_Ita_Catalog.txt} a random second decimal digit is added to ML and Md of the INGV
bulletin (ISIDe) before computing Mw proxies (Lolli et al., Seism. Res. Lett., 91,
3208-3222, doi: 10.1785/0220200148), in \code{HORUS_Ita_Catalog_o.txt} the original ML and Md
are used to compute Mw proxies. \code{ETAS.inlabru} includes a reformatted version
of the data from \code{HORUS_Ita_Catalog.txt} as a \code{data.frame}.

The data are provided as they are, no express or implied warranty is given.
}
\references{
Barbara Lolli(1), Daniele Randazzo(1), Gianfranco Vannucci(1) and Paolo Gasperini (2),(1)
(2020). The Homogenized Instrumental Seismic Catalog (HORUS) of Italy from 1960 to Present,
Seismol. Res. Lett, doi: 10.1785/0220200148.

(1) Istituto Nazionale di Geofisica e Vulcanologia, Sezione di Bologna
(2) Dipartimento di Fisica e Astronomia, Università di Bologna
}
