% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETAS.triggering.functions.R
\name{Int.ETAS.time.trig.function}
\alias{Int.ETAS.time.trig.function}
\title{Integrated ETAS time-triggering function}
\usage{
Int.ETAS.time.trig.function(theta, th, T2)
}
\arguments{
\item{theta}{ETAS parameters data.frame(mu=mu, K=K, alpha=alpha, c=c, p=p)}

\item{th}{Time of past event? \link{days}}

\item{T2}{End of temporal model domain.}
}
\description{
Integrated ETAS time-triggering function
}
