% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InjectionCode.R
\name{IntInjecIntensity}
\alias{IntInjecIntensity}
\title{Injection Rate function}
\usage{
IntInjecIntensity(a = 50, V.i = 1, tau = 10, T.i, T2)
}
\arguments{
\item{a}{Event rate per unit volume injected}

\item{V.i}{Injected volume}

\item{tau}{Decay rate \verb{[days]}}

\item{T.i}{Time of injection event}

\item{T2}{End of temporal model domain}
}
\value{
Forward time integrated function for exponential rate decay
}
\description{
Forward time integrated function for exponential rate decay
}
