% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InjectionCode.R
\name{IntInjectionIntensity}
\alias{IntInjectionIntensity}
\alias{Inv_IntInjectionIntensity}
\title{Injection Rate function calculations}
\usage{
IntInjectionIntensity(a = 50, V.i = 1, tau = 10, T.i, T2)

Inv_IntInjectionIntensity(
  a = 50,
  V.i = 1,
  tau = 10,
  T.i,
  number.injected.events
)
}
\arguments{
\item{a}{Event rate per unit volume injected}

\item{V.i}{Injected volume}

\item{tau}{Decay rate \verb{[days]}}

\item{T.i}{Time of injection event}

\item{T2}{End of temporal model domain}

\item{number.injected.events}{The number of expected injected events, used for the
inverse.}
}
\value{
\code{IntInjectionIntensity} returns the forward time integrated function
for exponential rate decay.

\code{Inv_IntInjectionIntensity} returns the end time corresponding to
a given expected number of injected events.
}
\description{
Forward time integrated function for exponential rate decay, and its inverse
}
