% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaTransformations.R
\name{inv.exp.t}
\alias{inv.exp.t}
\title{Copula transformation from an Exponential to a standard Normal distribution}
\usage{
inv.exp.t(x, rate)
}
\arguments{
\item{x}{values from an Exponential distribution, \code{vector}.}

\item{rate}{rate of the Exponential distribution, \code{scalar}.}
}
\value{
values from a standard Normal distribution, \code{vector} same length as \code{x}
}
\description{
Copula transformation from an Exponential to a standard Normal distribution
}
