% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETAS.triggering.functions.R
\name{Inv.Int.ETAS.time.trig.function}
\alias{Inv.Int.ETAS.time.trig.function}
\title{Inverse of integrated ETAS time-triggering function}
\usage{
Inv.Int.ETAS.time.trig.function(theta, omega, th)
}
\arguments{
\item{th}{}
}
\description{
Inverse of integrated ETAS time-triggering function
}
