% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InjectionCode.R
\name{Inv.IntInjecIntensity}
\alias{Inv.IntInjecIntensity}
\title{Returns end time given a ...}
\usage{
Inv.IntInjecIntensity(a = 50, V.i = 1, tau = 10, T.i, number.injected.events)
}
\arguments{
\item{a}{Event rate per unit volume injected}

\item{V.i}{Injected volume}

\item{tau}{Decay rate \verb{[days]}}

\item{T.i}{Time of injection event}

\item{number.injected.events}{TODO}
}
\description{
Returns end time given a ...
}
