% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaTransformations.R
\name{inv.unif.t}
\alias{inv.unif.t}
\title{Copula transformation from an Uniform to a standard Normal distribution}
\usage{
inv.unif.t(x, a, b)
}
\arguments{
\item{x}{values from an Uniform distribution, \code{vector}.}

\item{a}{minimum of the Uniform distribution, \code{scalar}.}

\item{b}{maximum of the Uniform distribution, \code{scalar}.}
}
\value{
values from a standard Normal distribution, \code{vector} same length as \code{x}
}
\description{
Copula transformation from an Uniform to a standard Normal distribution
}
