% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalBinning.R
\name{It_df}
\alias{It_df}
\title{Function to calculate the integral of Omori's law}
\usage{
It_df(param_, time.df)
}
\arguments{
\item{param_}{ETAS parameters vector (\eqn{\mu, K, \alpha, c, p}), only \eqn{c, p} are used.}

\item{time.df}{output of the function \code{\link[=time_grid]{time_grid()}}}
}
\value{
\code{vector} of integral values between each bin provided in \code{time.df}
}
\description{
Function to calculate the integral of Omori's law
}
