% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETAS.triggering.functions.R
\name{lambda_2}
\alias{lambda_2}
\title{conditional intensity (used by Inlabru)}
\usage{
lambda_2(th, t, ti.v, mi.v, M0)
}
\arguments{
\item{th}{Set of trial ETAS parameters ??}

\item{M0}{Minimum magnitude threshold}
}
\description{
conditional intensity (used by Inlabru)
}
