% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.ETAS.Triggering.Functions.R
\name{lambda_be}
\alias{lambda_be}
\title{Conditional intensity - used by bayesianETAS
Used for comparing output of inlabru with Bayesian ETAS}
\usage{
lambda_be(th, t, ti.v, mi.v, M0)
}
\arguments{
\item{M0}{Minimum magnitude threshold}
}
\description{
Conditional intensity - used by bayesianETAS
Used for comparing output of inlabru with Bayesian ETAS
}
