% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETAS.triggering.functions.R
\name{log.Lambda.h}
\alias{log.Lambda.h}
\title{Logarithm of the integral of the ETAS triggering function}
\usage{
\method{log}{Lambda.h}(theta, th, mh, M0, T1, T2)
}
\arguments{
\item{theta}{ETAS parameters \code{data.frame(mu=mu, K=K, alpha=alpha, c=c, p=p)}.}

\item{th}{Time of parent event.}

\item{mh}{Magnitude of parent event}

\item{M0}{Minimum magnitude threshold}

\item{T1}{Start of temporal model domain.}

\item{T2}{End of temporal model domain.}
}
\value{
Logarithm of the integral of the ETAS triggering function
}
\description{
Logarithm of the integral of the ETAS triggering function
}
