% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian.ETAS.Triggering.Functions.R
\name{log.Lambda_h}
\alias{log.Lambda_h}
\title{integrated triggering function - used by bayesianETAS
Used for comparing output of inlabru with Bayesian ETAS}
\usage{
\method{log}{Lambda_h}(th, ti, mi, M0, T1, T2)
}
\arguments{
\item{M0}{Minimum magnitude threshold}

\item{T1}{Start of temporal model domain.}

\item{T2}{End of temporal model domain.}
}
\description{
integrated triggering function - used by bayesianETAS
Used for comparing output of inlabru with Bayesian ETAS
}
