% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETAS.triggering.functions.R
\name{log.Lambda_h2}
\alias{log.Lambda_h2}
\title{integrated triggering function - used by Inlabru ALSO USED IN GENERATION SAMPLES}
\usage{
\method{log}{Lambda_h2}(theta, ti, mi, M0, T1, T2)
}
\arguments{
\item{theta}{ETAS parameters \code{data.frame(mu=mu, K=K, alpha=alpha, c=c, p=p)}.}

\item{ti}{Time of parent event.}

\item{mi}{Magnitude of parent event}

\item{M0}{Minimum magnitude threshold}

\item{T1}{Start of temporal model domain.}

\item{T2}{End of temporal model domain.}
}
\description{
integrated triggering function - used by Inlabru ALSO USED IN GENERATION SAMPLES
}
