% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETAS.triggering.functions.R
\name{loggaus.t}
\alias{loggaus.t}
\title{Log-gaussian copula transformation: Conversion of ETAS para to internal scale}
\usage{
loggaus.t(x, m, s)
}
\arguments{
\item{s}{}
}
\description{
Log-gaussian copula transformation: Conversion of ETAS para to internal scale
}
