% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticCatalogues.R
\name{sample.temoral.ETAS.daughters}
\alias{sample.temoral.ETAS.daughters}
\title{Generate a sample of new events \code{data.frame(t_i, M_i)} of length \code{n.ev} for parent event occuring at time \code{t_h} using the ETAS model}
\usage{
sample.temoral.ETAS.daughters(theta, beta.p, th, n.ev, M0, T1, T2)
}
\arguments{
\item{n.ev}{The number of events to be placed}

\item{M0}{Minimum magnitude in synthetic catalogue}

\item{T1}{Start time for synthetic catalogue \link{days}}

\item{T2}{End time for synthetic catalogue \link{days}}
}
\description{
Generate a sample of new events \code{data.frame(t_i, M_i)} of length \code{n.ev} for parent event occuring at time \code{t_h} using the ETAS model
}
