% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticCatalogues.R
\name{sample.temporal.ETAS.generation}
\alias{sample.temporal.ETAS.generation}
\title{Take all previous parent events from \code{Ht=data.frame[ts, magnitudes]} and generates their daughters events using the ETAS model}
\usage{
sample.temporal.ETAS.generation(theta, beta.p, Ht, M0, T1, T2, ncore = 1)
}
\arguments{
\item{theta}{ETAS parameters \code{data.frame(mu=mu, K=K, alpha=alpha, c=c, p=p)}.}

\item{beta.p}{Slope of GR relation: beta = b ln(10).}

\item{Ht}{The set of parent events in the form \code{data.frame[ts, magnitudes]}}

\item{M0}{The minimum earthquake magnitude in the synthetic catalogue.}

\item{T1}{The start time for the synthetic catalogue \link{days}.}

\item{T2}{The end time for the synthetic catalogue \link{days}.}

\item{ncore}{The number of compute cores to use}
}
\value{
Return one generation of daughters from the parents in \code{Ht} in the form \code{data.frame(t_i, M_i)}.
}
\description{
Take all previous parent events from \code{Ht=data.frame[ts, magnitudes]} and generates their daughters events using the ETAS model
}
