% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticCatalogues.R
\name{sample.temporal.ETAS.times}
\alias{sample.temporal.ETAS.times}
\title{Sampling times according to the EYAS triggering function}
\usage{
sample.temporal.ETAS.times(theta, n.ev, th, T2)
}
\arguments{
\item{theta}{The ETAS parameter set which}

\item{n.ev}{Number of events to return in the sample}

\item{T2}{}
}
\description{
Sampling times according to the EYAS triggering function
}
