% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticCatalogues.R
\name{sample_temporal_ETAS_daughters}
\alias{sample_temporal_ETAS_daughters}
\title{Generate a sample of new events \code{data.frame(t_i, M_i)} of length \code{n.ev} for one parent event occuring at time \code{t_h} using the ETAS model.}
\usage{
sample_temporal_ETAS_daughters(theta, beta.p, th, n.ev, M0, T1, T2)
}
\arguments{
\item{theta}{ETAS parameters \code{list(mu=mu, K=K, alpha=alpha, c=c, p=p)}.}

\item{beta.p}{Slope of GR relation: beta = b ln(10).}

\item{th}{Time of parent event \verb{[days]}.}

\item{n.ev}{The number of events to be placed.}

\item{M0}{Minimum magnitude in synthetic catalogue.}

\item{T1}{Start time for synthetic catalogue \verb{[days]}.}

\item{T2}{End time for synthetic catalogue \verb{[days]}.}
}
\value{
Generate a sample of new events \code{data.frame(t_i, M_i)} from one parent
}
\description{
Generate a sample of new events \code{data.frame(t_i, M_i)} of length \code{n.ev} for one parent event occuring at time \code{t_h} using the ETAS model.
}
