% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticCatalogues.R
\name{sample_temporal_ETAS_generation}
\alias{sample_temporal_ETAS_generation}
\title{Take all previous parent events from \code{Ht=data.frame[ts, magnitudes]} and generates their daughters events using the ETAS model}
\usage{
sample_temporal_ETAS_generation(theta, beta.p, Ht, M0, T1, T2, ncore = NULL)
}
\arguments{
\item{theta}{ETAS parameters \code{list(mu=mu, K=K, alpha=alpha, c=c, p=p)}.}

\item{beta.p}{Slope of GR relation: beta = b ln(10).}

\item{Ht}{The set of parent events in the form \code{data.frame[ts, magnitudes]}}

\item{M0}{The minimum earthquake magnitude in the synthetic catalogue.}

\item{T1}{The start time for the synthetic catalogue \verb{[days]}.}

\item{T2}{The end time for the synthetic catalogue \verb{[days]}.}

\item{ncore}{Deprecated argument for controlling parallelism. Use
\code{future::plan(future::multisession, workers = ncore)} (or similar) to configure
parallelism in your code instead.}
}
\value{
Return one generation of daughters from the parents in \code{Ht} in the form \code{data.frame(t_i, M_i)}.
}
\description{
Take all previous parent events from \code{Ht=data.frame[ts, magnitudes]} and generates their daughters events using the ETAS model
}
\examples{
# The parents are specified in Ht
Ht <- data.frame(ts = c(500), magnitudes = c(6.7))
sample_temporal_ETAS_generation(
  theta = list(mu = 0.1, K = 0.089, alpha = 2.29, c = 0.11, p = 1.08),
  beta.p = log(10),
  M0 = 2.5,
  T1 = 0, T2 = 1000,
  Ht = Ht
)
}
