% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticCatalogues.R
\name{sample_temporal_ETAS_times}
\alias{sample_temporal_ETAS_times}
\title{Sample times for events triggered by a parent according to the ETAS
triggering function}
\usage{
sample_temporal_ETAS_times(theta, n.ev, th, T2)
}
\arguments{
\item{theta}{ETAS parameters \code{list(mu=mu, K=K, alpha=alpha, c=c, p=p)}.}

\item{n.ev}{Number of events to return in the sample in time domain \verb{(th, T2]}.}

\item{th}{Time of the parent event producing \code{n.ev} daughters.}

\item{T2}{End time of model domain.}
}
\value{
t.sample A list of times in the interval \verb{[0, T2]} distributed
according to the ETAS triggering function.
}
\description{
Sampling times for events triggered by a parent at th according to the ETAS
triggering function
}
