% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HawkesProcess.R
\name{Temporal.ETAS.fit}
\alias{Temporal.ETAS.fit}
\title{Fits the remporal ETAS model and returns the results. This function decomposes the input.list for the `Hawkes.bru2`` function.}
\usage{
Temporal.ETAS.fit(input.list)
}
\arguments{
\item{input.list}{All input data and parameters are passed to inlabru via this structured list.}
}
\value{
The inlabru Hawkes process results data.frame
}
\description{
Fits the remporal ETAS model and returns the results. This function decomposes the input.list for the `Hawkes.bru2`` function.
}
