% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HawkesProcess.R
\name{Temporal.ETAS}
\alias{Temporal.ETAS}
\title{Function to fit Hawkes process model}
\usage{
Temporal.ETAS(
  total.data,
  M0,
  T1,
  T2,
  link.functions = NULL,
  coef.t.,
  delta.t.,
  N.max.,
  bru.opt
)
}
\arguments{
\item{total.data}{Observed events: \code{data.frame} with columns time (ts), magnitude (magnitudes), event identifier (idx.p). Column names must not be changed.}

\item{M0}{Minimum magnitude threshold, \code{scalar}}

\item{T1}{Start of temporal model domain, \code{scalar} \verb{[measure unit of sample.s$ts]}.}

\item{T2}{End of temporal model domain, \code{scalar} \verb{[measure unit of sample.s$ts]}.}

\item{link.functions}{Functions to transform the parameters from the internal INLA scale to the ETAS scale. It must be a \code{list} of functions with names (mu, K, alpha, c_, p)}

\item{coef.t.}{TimeBinning parameter: parameter regulating the relative length of successive bins, \code{scalar}.}

\item{delta.t.}{TimeBinning parameter: parameter regulating the bins' width, \code{scalar}.}

\item{N.max.}{TimeBinning parameter: parameter regulating the Number of bins (= \code{N.max} + 2), \code{scalar}.}

\item{bru.opt}{Runtime options for inlabru: See https://inlabru-org.github.io/inlabru/reference/bru_call_options.html, \code{list}}
}
\value{
The fitted model as a 'bru' object, which is a list
}
\description{
function to fit a temporal ETAS model using \code{inlabru}.
}
