% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalBinning.R
\name{time_grid}
\alias{time_grid}
\title{Generate a set of time bins for a specific event.}
\usage{
time_grid(data.point, coef.t, delta.t, N.exp., T1., T2.)
}
\arguments{
\item{data.point}{Point for which the binning is calculated, \code{list} with
elements time (\verb{ts, scalar}), event index (\verb{idx.p, scalar}). Names are
mandatory and should not be changed}

\item{coef.t}{TimeBinning parameter: look \code{\link[=breaks_exp]{breaks_exp()}}}

\item{delta.t}{TimeBinning parameter: look \code{\link[=breaks_exp]{breaks_exp()}}}

\item{N.exp.}{TimeBinning parameter: look \code{\link[=breaks_exp]{breaks_exp()}}}

\item{T1.}{Start of the temporal domain, \code{scalar}}

\item{T2.}{End of the temporal domain \code{scalar}.}
}
\value{
A \code{data.frame} with as many rows as the number of bins and fixed number of columns.
The columns are
\itemize{
\item \code{t.start} : starting point of the bin (minimum = \code{T1.}).
\item \code{t.end} : end point of the bin. (maximum = \code{T2.}).
\item \code{t.bin.name} : unique bin identifier.
\item \code{t.ref_layer} : bin identifier for calculations
\item \code{ts} : time provided in \code{data.point}
\item \code{idx.p} : identifier provided in \code{data.point}
}
The bins are only between \code{T1.} and \code{T2.} or containing \code{T1.}
}
\description{
Generate a set of time bins for a specific event.
}
\examples{
## EXAMPLE 1
event <- list(ts = 0, idx.p = 1)
time_grid(data.point = event, coef.t = 1, delta.t = 0.1, N.exp. = 8, T1. = 1, T2. = 20)
}
