% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{triggering_fun_plot}
\alias{triggering_fun_plot}
\title{Triggering function plot from posterior samples}
\usage{
triggering_fun_plot(
  input.list,
  post.samp = NULL,
  n.samp = 10,
  magnitude = 4,
  t.end = 1,
  n.breaks = 100
)
}
\arguments{
\item{input.list}{structured input \code{list} with at least two elements:
\itemize{
\item \code{model.fit}: \code{bru} object used to sample the posterior of the ETAS
parameters
\item \code{link.functions}: \code{list} of functions to convert the ETAS parameters
from the INLA scale to the ETAS scale
}}

\item{post.samp}{\code{data.frame} containing posterior samples of the parameters.
If \code{NULL}, then \code{n.samp} samples are generated. If \code{n.samp} is different
from \code{nrow(post.samp)} then \code{n.samp} rows are uniformly sampled from
\code{post.samp}. Default is \code{NULL}}

\item{n.samp}{Number of posterior samples, \code{integer} (\code{default = 10}).}

\item{magnitude}{Magnitude of the event for which the triggering function is
calculated, \code{scalar} (\code{default = 4}).}

\item{t.end}{Upper bound of the x-axis, \code{scalar} (\code{default = 1}).}

\item{n.breaks}{Number of points between 0 and \code{t.end} to calculate the
function, \code{integer} (\code{default = 100})}
}
\value{
\code{ggplot} object with grey lines representing the triggering function
for each posterior sample. Black lines representing the 0.025 and 0.975
quantiles of the function values calculated for each posterior sample.
Horizontal red lines represents the 0.025 and 0.975 quantiles of the
sampled background rates.
}
\description{
Function to plot the ETAS triggering function corresponding to different
posterior samples
}
