% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETAS.triggering.functions.R
\name{unif.t}
\alias{unif.t}
\title{Uniform copula transformation: Conversion of ETAS para to internal scale}
\usage{
unif.t(x, a, b)
}
\arguments{
\item{b}{}
}
\description{
Uniform copula transformation: Conversion of ETAS para to internal scale
}
