% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupInlabruInputs.R
\name{create.input.list.temporal.withCatalogue}
\alias{create.input.list.temporal.withCatalogue}
\title{Function to create a default input file for the ETAS Hawkes temporal model where a catalogue is specified in the input file.}
\usage{
create.input.list.temporal.withCatalogue(input_path)
}
\arguments{
\item{input_path}{}
}
\description{
Function to create a default input file for the ETAS Hawkes temporal model where a catalogue is specified in the input file.
}
