% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupInlabruInputs.R
\name{create_input_list_temporal_noCatalogue}
\alias{create_input_list_temporal_noCatalogue}
\title{Create input list for ETAS Hawkes temporal model without catalogue}
\usage{
create_input_list_temporal_noCatalogue(input_path, num.threads = NULL)
}
\arguments{
\item{input_path}{Input file and path as a string}

\item{num.threads}{Optional argument for the number of threads to be used by
parallel processing by inlabru/INLA}
}
\value{
The formatted input.list with the elements required for the temporal
Hawkes model
}
\description{
Create input list for ETAS Hawkes temporal model without catalogue
}
\examples{
create_input_list_temporal_noCatalogue(
  system.file(
    "extdata",
    "user_input_synthetic_noCatalogue.txt",
    package = "ETAS.inlabru"
  )
)
}
