% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETAS.inlabru-package.R
\docType{package}
\name{ETAS.inlabru-package}
\alias{ETAS.inlabru}
\alias{ETAS.inlabru-package}
\title{ETAS.inlabru: Bayesian ETAS model for modelling seismic sequences with inlabru}
\description{
Modelling and inversion of ETAS model of seismicity using inlabru The Epidemic Type Aftershock Sequence (ETAS) model is designed to model earthquakes that are triggered by previous events. In statistics, this is referred to as a Hawkes process. The code can be used to generate synthetic ETAS catalogues which can also include some seeded events to model specific sequences. We also implement a Bayesian inversion scheme using the Integrated Nested Laplace Approximation (INLA) using inlabru. For the temporal model, given a training catalogue of times and magnitudes, the code returns the joint posteriors for all the ETAS parameters. In the future roadmap, we will include tools to model the spatial distribution and spatio-temporal evolution of seismic sequences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://edinburgh-seismicity-hub.github.io/ETAS.inlabru/}
  \item \url{https://github.com/edinburgh-seismicity-hub/ETAS.inlabru}
  \item Report bugs at \url{https://github.com/edinburgh-seismicity-hub/ETAS.inlabru/issues}
}

}
\author{
\strong{Maintainer}: Francesco Serafini \email{francesco.serafini@newcastle.ac.uk} (\href{https://orcid.org/0000-0003-0154-6200}{ORCID})

Authors:
\itemize{
  \item Mark Naylor \email{mark.naylor@ed.ac.uk} (\href{https://orcid.org/0000-0002-3761-5522}{ORCID}) [thesis advisor]
  \item Finn Lindgren \email{Finn.Lindgren@ed.ac.uk} (\href{https://orcid.org/0000-0002-5833-2011}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
