% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaTransformations.R
\name{gamma_t}
\alias{gamma_t}
\title{Copula transformation from a standard Normal distribution to a Gamma distribution}
\usage{
gamma_t(x, a, b)
}
\arguments{
\item{x}{values from a standard Normal distribution, \code{vector}.}

\item{a}{shape parameter of the gamma distribution \code{scalar}.}

\item{b}{rate parameter of the gamma distribution \code{scalar}.}
}
\value{
values from a Gamma distribution with shape \code{a} and rate \code{b}, \code{vector} same length as \code{x}.
}
\description{
Copula transformation from a standard Normal distribution to a Gamma distribution
}
