% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysingPosteriors.R
\name{get_posterior_N}
\alias{get_posterior_N}
\title{Plot the posterior distribution of the expected number of events}
\usage{
get_posterior_N(input.list)
}
\arguments{
\item{input.list}{Which has combined the input file (for link functions) and bru output (for marginals)}
}
\value{
A \code{list} of three objects:
\itemize{
\item \code{post.df}: \code{data.frame} containing posterior informations on the posterior distribution of the number of events
\item \code{post.plot} : \code{ggplot} object showing the posterior distribution of the expected number of events
\item \code{post.plot.shaded} : \code{ggplot} object showing the posterior distribution of the expected number of events, shaded region corresponds to the 0.025 and 0.975 quantiles of the distribution of the distribution of the number of events.}
}
\description{
Plot the posterior distribution of the expected number of events
}
