% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSyntheticCatalogues.R
\name{Int.ETAS.time.trig.function}
\alias{Int.ETAS.time.trig.function}
\title{Integrated Omori's law}
\usage{
Int.ETAS.time.trig.function(theta, th, T2)
}
\arguments{
\item{theta}{ETAS parameters \code{list(mu=mu, K=K, alpha=alpha, c=c, p=p)}}

\item{th}{Time of past event \link{days} and start of temporal domain, \code{vector}.}

\item{T2}{End of temporal domain, \code{scalar}.}
}
\value{
Value of the integral of the Omori's law
}
\description{
Integrated Omori's law
}
\details{
The function returns the integral of the Omori's law, namely
\deqn{\int_{t_h}^{T_2} \left(\frac{t - t_h}{c} + 1\right)^{-p} dt}
}
