% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InjectionCode.R
\name{IntInjecIntensity}
\alias{IntInjecIntensity}
\title{Injection Rate function}
\usage{
IntInjecIntensity(a = 50, V.i = 1, tau = 10, T.i, T2)
}
\arguments{
\item{a}{Event rate per unit volume injected}

\item{V.i}{Injected volume}

\item{tau}{Decau rate \link{days}}

\item{T.i}{Time of injection event}

\item{T2}{}
}
\description{
Injection Rate function
}
