% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaTransformations.R
\name{loggaus_t}
\alias{loggaus_t}
\title{Copula transformation from a standard Normal distribution to a Log-Normal distribution}
\usage{
loggaus_t(x, m, s)
}
\arguments{
\item{x}{values from a standard Normal distribution, \code{vector}.}

\item{m}{mean of the logarithm of the Log-Normal distribution, \code{scalar}.}

\item{s}{standard deviation of the logarithm of the Log-Normal distribution, \code{scalar}.}
}
\value{
Values from a Log-Normal distribution with logarithmic mean \code{m} and standard deviation \code{s}, \code{vector} same length as \code{x}.
}
\description{
Copula transformation from a standard Normal distribution to a Log-Normal distribution
}
\seealso{
Other copula-transformations: 
\code{\link{exp_t}()},
\code{\link{gamma_t}()},
\code{\link{inv_exp_t}()},
\code{\link{inv_gamma_t}()},
\code{\link{inv_loggaus_t}()},
\code{\link{inv_unif_t}()},
\code{\link{unif_t}()}
}
\concept{copula-transformations}
